Use Flow's catch() operator instead of try-catch
https://kotlinlang.org/docs/reference/coroutines/flow.html#exception-transparency. According to this, 
Flows must be transparent to exceptions and it is a violation of the exception transparency to emit values in the flow { ... } builder from inside of a try/catch block. This guarantees that a collector throwing an exception can always catch it using try/catch as in the previous example. 

I think it's better to use Kotlin Flow's built-in catch operator. This won't affect anything according to the current code but it will be helpful if there is any change in the catch() block in the future.